add utils tests
